Destroy bug fixed

+++ base/app/models/activity.rb
@@ -410,7 +410,9 @@ class Activity < ActiveRecord::Base
   # Destroy children comments when the activity is destroyed
   def destroy_children_comments
     comments.each do |c|
-      c.direct_object.destroy
+      unless c.direct_object.nil?
+        c.direct_object.destroy
+      end
     end
   end
